requirements.txt
customtkinter>=5.2.0
pyyaml>=6.0
polib>=1.2.0
